+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 Expect active development and potentially significant breaking changes in the `0.x` track. We'll try to be diligent about releasing a `1.0` version in a timely fashion (ideally within 3 to 6 months), to signal the start of a more stable API.
 
-### vNEXT
+### v0.4.20
 - Fix: Warn but do not fail when refetchQueries includes an unknown query name [PR #700](https://github.com/apollostack/apollo-client/pull/700)
 - Fix: avoid field error on mutations after a query cancellation or a query failure by enforcing returnPartialData during previous data retrieval before applying a mutation update. [PR #696](https://github.com/apollostack/apollo-client/pull/696) and [Issue #647](https://github.com/apollostack/apollo-client/issues/647).
+- Add observableQuery.setVariables function [PR #635](https://github.com/apollostack/apollo-client/pull/635)
+- Add observableQuery.currentResult function [PR #697](https://github.com/apollostack/apollo-client/pull/697)
+- Update to typescript 2.0.3 [PR #697](https://github.com/apollostack/apollo-client/pull/697)
 
 ### v0.4.19
 - Fix: set default reduxRootKey for backwards-compatibility when using ApolloClient as middleware  [PR #688](https://github.com/apollostack/apollo-client/pull/688)